fixed files

+++ src/com/bc/DisplayFunctions.java
@@ -21,6 +21,9 @@ public class DisplayFunctions {
 //	6. The Fees 7. The Taxes 8. The Total for the Invoice 9. The totals of all theses as a final line
 	public static void summaryReport(List<Invoice> lInv, List<Customer> lc, ArrayList<Person> lpers,
 
+			
+			
+			
 			List<Product> lprod) throws IOException {
 		Double loyaltyDiscount = 0.0;
 
@@ -39,7 +42,8 @@ public class DisplayFunctions {
 
 		for (Invoice s : lInv) {
 
-//Check Owner code
+
+			//Check Owner code
 
 			String ownerCode = s.getOwnerCode();
 
@@ -257,7 +261,7 @@ public class DisplayFunctions {
 
 			Double partsCost = 0.0;
 
-			Integer concessionUnits = 0;
+			Double concessionUnits = 0.0;
 
 			Double itemTax = 0.0;
 
@@ -368,8 +372,8 @@ public class DisplayFunctions {
 //						Concessions
 						else if (p.getProductType().equals("C")) {
 							productInfo = "units";
-							Concession c = new Concession((Concession) p, Integer.parseInt(itemTokens[1]));
-							c.setQuantity(Integer.parseInt(itemTokens[1]));
+							Concession c = new Concession((Concession) p, Double.parseDouble(itemTokens[1]));
+							c.setQuantity(Double.parseDouble(itemTokens[1]));
 							productRate = c.getUnitCost();
 							productSubtotal = c.getConcessionCost();
 							extraInfo = "";
@@ -517,8 +521,8 @@ public class DisplayFunctions {
 							lrep.add(f);
 
 						} else if (p.getProductType().equals("C")) {
-							Concession c = new Concession((Concession) p, Integer.parseInt(itemTokens[1]));
-							c.setQuantity(Integer.parseInt(itemTokens[1]));
+							Concession c = new Concession((Concession) p, Double.parseDouble(itemTokens[1]));
+							c.setQuantity(Double.parseDouble(itemTokens[1]));
 							itemCost = c.getConcessionCost();
 							subTotal += itemCost;
 						}
@@ -531,8 +535,8 @@ public class DisplayFunctions {
 				for (Product p : productList) {
 					if (p.getProductCode().equals(itemTokens[0])) {
 
-						Concession c = new Concession((Concession) p, Integer.parseInt(itemTokens[1]), itemTokens[2]);
-						c.setQuantity(Integer.parseInt(itemTokens[1]));
+						Concession c = new Concession((Concession) p, Double.parseDouble(itemTokens[1]), itemTokens[2]);
+						c.setQuantity(Double.parseDouble(itemTokens[1]));
 
 						itemCost = c.getConcessionCost();
 						subTotal += itemCost;
